refactor: rename response-builder file to wallet-response
