+++ packages/server-wallet/src/protocols/close-channel.ts
@@ -6,7 +6,7 @@ import {Store} from '../wallet/store';
 import {Bytes32} from '../type-aliases';
 import {ChainServiceInterface} from '../chain-service';
 import {DBCloseChannelObjective} from '../models/objective';
-import {WalletResponse} from '../wallet/response-builder';
+import {WalletResponse} from '../wallet/wallet-response';
 import {recordFunctionMetrics} from '../metrics';
 import {Channel} from '../models/channel';
 import {LedgerRequest} from '../models/ledger-request';
